% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_lactCurve.R
\name{calc_coeff}
\alias{calc_coeff}
\title{Calculate coefficients to model lactation curve of (mountain-pastured) dairy cows}
\usage{
calc_coeff(model, dataInput, y_field, t_field, w_field = NULL,
  t1_field = NULL, diff_field = NULL, fullInteraction = FALSE,
  k_wilmink = 0.1, diff_value = NULL, endCurve = FALSE,
  alpDuration = 115)
}
\arguments{
\item{model}{char One of 'AS' (for Ali-Schaeffer), 'Wilmink','Wood' to describe the type of regression}

\item{dataInput}{dataframe The dataframe containing the observations in rows with the columns as givenin y_field, t_field etc...}

\item{y_field}{char The name of the column in \code{dataInput} that corresponds to the response field (typically milk)}

\item{t_field}{char The name of the column in \code{dataInput} that corresponds to the time of observation (typically days in milk)}

\item{w_field}{char The name of the column in \code{dataInput} that corresponds to the weight of the observation (typically the number of cows when working on average of cows)}

\item{t1_field}{char The name of the column in \code{dataInput} that corresponds to the time (same unit as \code{t_field}) at which the time is alped. If groups of cows are taken, this time must be different for each calving season. If NULL, the base model as presented by AS/Wood/Wilmink will be returned}

\item{diff_field}{char The name of the column in \code{dataInput} that corresponds to the field for which we want the alp term (or all terms if \code{fullInteraction} is TRUE) to vary}

\item{fullInteraction}{boolean If FALSE only the alp term will vary according to \code{diff_field}. If TRUE all terms are allowed to vary according to \code{g_field}.}

\item{k_wilmink}{real If model='Wilmink', the value of the k coefficient. By defautl, 0.1.}

\item{diff_value}{char/integer vector The possible value that \code{diff_field can have}. The first value will be taken as the reference value}

\item{endCurve}{boolean Whether the end of the curve (after high alpine grazing is over) should be plotted or not}

\item{alpDuration}{integer The duration of the high alpine grazing season (same unit as \code{t_field})}
}
\value{
An obeject of class "lm", as the result of the lm() call
}
\description{
Calculate coefficients to model lactation curve of (mountain-pastured) dairy cows
}
\details{
The modelling is based on three types of possible base model (see \code{model}), and adaptation is done for mountain-pastured cows: 1) increased linear decrease during high alpine grazing 2) sharp increase after the transhumance is over 3) smoother decrease for the end of the lactation
}
\examples{

}
\author{
Solange Duruz
}
