% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_lactCurve.R
\name{plot_lc}
\alias{plot_lc}
\title{Plot, model and test the significance of parameters for lactation curve modelling of (mountain-pastured cows)}
\usage{
plot_lc(dataInput, y_field, t_field, cm_field, month = c(12, 11, 10, 9,
  8, 3, 2, 1), t1_field, diff_field = NULL, diff_value = c(1, 0),
  w_field = NULL, prediction = TRUE, model = "Wilmink",
  k_wilmink = 0.1, endCurve = FALSE, alpDuration = 115,
  fullInteraction = FALSE, interactionMonth = FALSE,
  ylabel = "Milk yield [kg]", xlabel = "Days in milk", pal = "Blues",
  predictionCol = "gray", pch = c(1, 19, 3), diffLegend = NULL)
}
\arguments{
\item{dataInput}{a matrix containing lines observations at different days in milk.}

\item{y_field}{char The name of the column in \code{dataInput} containing the y-observation (typically milk yield but could also be protein yield)}

\item{t_field}{char The name of the column in \code{dataInput}  containing the time of the observation (typically in days in milk)}

\item{cm_field}{char The name of the column in \code{dataInput}  containing the season of calving (if different calving month are present in the dataset)}

\item{month}{char/integer vector The list of months to consider as defined in \code{cm_field}}

\item{t1_field}{char The name of the column in \code{dataInput} that corresponds to the time (same unit as \code{t_field}) at which the time is alped. If groups of cows are taken, this time must be different for each calving season. If NULL, the base model as presented by AS/Wood/Wilmink will be returned}

\item{diff_field}{char The name of the column in \code{dataInput} containing the groups (if want to test the difference among groups)}

\item{diff_value}{char/integer vector The possible value that \code{diff_field can have}. The first value will be taken as the reference value}

\item{w_field}{char The name of the column in \code{dataInput} that corresponds to the weight of the observation (typically the number of cows when working on average of cows)}

\item{prediction}{boolean Whether the prediction line should be plotted on top of the curve. Necessarily TRUE if want to calculate p-values of}

\item{model}{char One of 'AS' (for Ali-Schaeffer), 'Wilmink','Wood' to describe the type of regression}

\item{k_wilmink}{real If model='Wilmink', the value of the k coefficient. By defautl, 0.1.}

\item{endCurve}{boolean Whether the end of the curve (after high alpine grazing is over) should be plotted or not}

\item{alpDuration}{integer The duration of the high alpine grazing season (same unit as \code{t_field})}

\item{fullInteraction}{boolean If FALSE only the alp term will vary according to \code{diff_field}. If TRUE all terms are allowed to vary according to \code{g_field}.}

\item{interactionMonth}{boolean If true, the coefficients of the regression will be estimated separately for each month}

\item{ylabel}{char The name of the y-axis label in the plot}

\item{xlabel}{char The name of the x-axis label in the plot}

\item{pal}{char The name of the color palette used to draw the data points as defined in color brewer. See ?RColorBrewer::brewer.pal for mor details}

\item{predictionCol}{char The color name for the prediction line}

\item{pch}{integer vector The pch code for the different \code{diff_value} present in the dataset}

\item{diffLegend}{?}
}
\value{
A list containing the coefficients of the regression ($coeffs), the d-parameter ($alp_coeff), the estimated total milk production ($milk_total), and during alpine grazine only ($milk_alp). If \code{diff_field} is not null, the list will aslo contain  the change in the d-parameter according to the group ($alpdiff_coeff), the p-value of the difference between groups ($pval)
}
\description{
The modelling will be performed with the function \code{calc_coeff}
}
\examples{


}
\author{
Solange Duruz
}
